+++ app/src/main/java/com/example/mtot/retrofit2/RetrofitObject.kt
@@ -5,7 +5,6 @@ import com.google.gson.GsonBuilder
 import okhttp3.OkHttpClient
 import okhttp3.logging.HttpLoggingInterceptor
 import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
 import retrofit2.converter.scalars.ScalarsConverterFactory
 
 object LoginObject {
@@ -23,7 +22,7 @@ object LoginObject {
         .addConverterFactory(ScalarsConverterFactory.create())
         .client(okHttpClient())
         .build()
-    var loginInterface: LoginInterface = retrofit.create(LoginInterface::class.java)
+    var loginInterface: RetrofitInterface = retrofit.create(RetrofitInterface::class.java)
 
 }
