+++ app/src/main/java/com/example/mtot/ui/social/SocialFragment.kt
@@ -1,9 +1,7 @@
 package com.example.mtot.ui.social
 
-import android.app.AlertDialog
 import android.content.Intent
 import android.os.Bundle
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -11,12 +9,7 @@ import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.example.mtot.R
 import com.example.mtot.databinding.FragmentSocialBinding
-import com.example.mtot.retrofit2.FriendObject
 import com.example.mtot.retrofit2.FriendData
-import com.example.mtot.retrofit2.SharedPreference.saveFriendData
-import retrofit2.Call
-import retrofit2.Callback
-import retrofit2.Response
 
 class SocialFragment : Fragment() {
     lateinit var binding: FragmentSocialBinding
